Add files via upload
Hello, I coded the 3 cards for my website on Sublime Text but they do not appear... I coded the cards firstly by using "inspect" directly on Chrome like shown on the video. Then I copy it on Sublime Text and save it. But wthen I refresh the page all of the modifications I did for the cards are gone. I tried to fix this issue but it did not work. This is what I did on the card.css file.

.cards {
	display: flex;
	flex-wrap: wrap;
}

.card {
       display: flex;
       align-items: center;
        justify-content: space-between;
        padding: 20px;
	box-shadow: 0px 10px 30px rgba(0,0,0,0.2);
	border-radius: 6px;
	margin: 10px;
	flex-basis: 250px;
	flex-grow: 1;
}

.card img {
  width: 150px;
}
.card h3 {
  flex-grow: 1;
  padding-left: 20px;
  font-size: 24px;
}